requirements.txt
flask>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
Pillow>=9.0
inky[rpi]>=1.5
gpiozero>=1.6
requests>=2.28
APScheduler>=3.9
qrcode>=7.0
